Improved type error messages.
